+++ bmc-core/src/main/java/com/oracle/bmc/core/VirtualNetworkAsync.java
@@ -3393,7 +3393,8 @@ public interface VirtualNetworkAsync extends AutoCloseable {
                             handler);
 
     /**
-     * Lists the network security groups in the specified compartment.
+     * Lists either the network security groups in the specified compartment, or those associated with the specified VLAN.
+     * You must specify either a `vlanId` or a `compartmentId`, but not both. If you specify a `vlanId`, all other parameters are ignored.
      *
      *
      * @param request The request object containing the details to send