+++ bmc-database/src/main/java/com/oracle/bmc/database/model/CreateAutonomousDatabaseCloneDetails.java
@@ -87,6 +87,24 @@ public class CreateAutonomousDatabaseCloneDetails extends CreateAutonomousDataba
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("kmsKeyId")
+        private String kmsKeyId;
+
+        public Builder kmsKeyId(String kmsKeyId) {
+            this.kmsKeyId = kmsKeyId;
+            this.__explicitlySet__.add("kmsKeyId");
+            return this;
+        }
+
+        @com.fasterxml.jackson.annotation.JsonProperty("vaultId")
+        private String vaultId;
+
+        public Builder vaultId(String vaultId) {
+            this.vaultId = vaultId;
+            this.__explicitlySet__.add("vaultId");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("adminPassword")
         private String adminPassword;
 
@@ -291,6 +309,8 @@ public class CreateAutonomousDatabaseCloneDetails extends CreateAutonomousDataba
                             dbWorkload,
                             dataStorageSizeInTBs,
                             isFreeTier,
+                            kmsKeyId,
+                            vaultId,
                             adminPassword,
                             displayName,
                             licenseModel,
@@ -325,6 +345,8 @@ public class CreateAutonomousDatabaseCloneDetails extends CreateAutonomousDataba
                             .dbWorkload(o.getDbWorkload())
                             .dataStorageSizeInTBs(o.getDataStorageSizeInTBs())
                             .isFreeTier(o.getIsFreeTier())
+                            .kmsKeyId(o.getKmsKeyId())
+                            .vaultId(o.getVaultId())
                             .adminPassword(o.getAdminPassword())
                             .displayName(o.getDisplayName())
                             .licenseModel(o.getLicenseModel())
@@ -368,6 +390,8 @@ public class CreateAutonomousDatabaseCloneDetails extends CreateAutonomousDataba
             DbWorkload dbWorkload,
             Integer dataStorageSizeInTBs,
             Boolean isFreeTier,
+            String kmsKeyId,
+            String vaultId,
             String adminPassword,
             String displayName,
             LicenseModel licenseModel,
@@ -396,6 +420,8 @@ public class CreateAutonomousDatabaseCloneDetails extends CreateAutonomousDataba
                 dbWorkload,
                 dataStorageSizeInTBs,
                 isFreeTier,
+                kmsKeyId,
+                vaultId,
                 adminPassword,
                 displayName,
                 licenseModel,