+++ bmc-database/src/main/java/com/oracle/bmc/database/model/CreateDbHomeWithDbSystemIdDetails.java
@@ -88,6 +88,15 @@ public class CreateDbHomeWithDbSystemIdDetails extends CreateDbHomeBase {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("isDesupportedVersion")
+        private Boolean isDesupportedVersion;
+
+        public Builder isDesupportedVersion(Boolean isDesupportedVersion) {
+            this.isDesupportedVersion = isDesupportedVersion;
+            this.__explicitlySet__.add("isDesupportedVersion");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("dbSystemId")
         private String dbSystemId;
 
@@ -127,6 +136,7 @@ public class CreateDbHomeWithDbSystemIdDetails extends CreateDbHomeBase {
                             databaseSoftwareImageId,
                             freeformTags,
                             definedTags,
+                            isDesupportedVersion,
                             dbSystemId,
                             dbVersion,
                             database);
@@ -143,6 +153,7 @@ public class CreateDbHomeWithDbSystemIdDetails extends CreateDbHomeBase {
                             .databaseSoftwareImageId(o.getDatabaseSoftwareImageId())
                             .freeformTags(o.getFreeformTags())
                             .definedTags(o.getDefinedTags())
+                            .isDesupportedVersion(o.getIsDesupportedVersion())
                             .dbSystemId(o.getDbSystemId())
                             .dbVersion(o.getDbVersion())
                             .database(o.getDatabase());
@@ -167,6 +178,7 @@ public class CreateDbHomeWithDbSystemIdDetails extends CreateDbHomeBase {
             String databaseSoftwareImageId,
             java.util.Map<String, String> freeformTags,
             java.util.Map<String, java.util.Map<String, Object>> definedTags,
+            Boolean isDesupportedVersion,
             String dbSystemId,
             String dbVersion,
             CreateDatabaseDetails database) {
@@ -176,7 +188,8 @@ public class CreateDbHomeWithDbSystemIdDetails extends CreateDbHomeBase {
                 kmsKeyVersionId,
                 databaseSoftwareImageId,
                 freeformTags,
-                definedTags);
+                definedTags,
+                isDesupportedVersion);
         this.dbSystemId = dbSystemId;
         this.dbVersion = dbVersion;
         this.database = database;