+++ bmc-loganalytics/src/main/java/com/oracle/bmc/loganalytics/model/StreamAction.java
@@ -48,11 +48,21 @@ public class StreamAction extends Action {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("savedSearchDuration")
+        private String savedSearchDuration;
+
+        public Builder savedSearchDuration(String savedSearchDuration) {
+            this.savedSearchDuration = savedSearchDuration;
+            this.__explicitlySet__.add("savedSearchDuration");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonIgnore
         private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();
 
         public StreamAction build() {
-            StreamAction __instance__ = new StreamAction(savedSearchId, metricExtraction);
+            StreamAction __instance__ =
+                    new StreamAction(savedSearchId, metricExtraction, savedSearchDuration);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
             return __instance__;
         }
@@ -60,7 +70,9 @@ public class StreamAction extends Action {
         @com.fasterxml.jackson.annotation.JsonIgnore
         public Builder copy(StreamAction o) {
             Builder copiedBuilder =
-                    savedSearchId(o.getSavedSearchId()).metricExtraction(o.getMetricExtraction());
+                    savedSearchId(o.getSavedSearchId())
+                            .metricExtraction(o.getMetricExtraction())
+                            .savedSearchDuration(o.getSavedSearchDuration());
 
             copiedBuilder.__explicitlySet__.retainAll(o.__explicitlySet__);
             return copiedBuilder;
@@ -75,10 +87,12 @@ public class StreamAction extends Action {
     }
 
     @Deprecated
-    public StreamAction(String savedSearchId, MetricExtraction metricExtraction) {
+    public StreamAction(
+            String savedSearchId, MetricExtraction metricExtraction, String savedSearchDuration) {
         super();
         this.savedSearchId = savedSearchId;
         this.metricExtraction = metricExtraction;
+        this.savedSearchDuration = savedSearchDuration;
     }
 
     /**
@@ -89,6 +103,34 @@ public class StreamAction extends Action {
 
     @com.fasterxml.jackson.annotation.JsonProperty("metricExtraction")
     MetricExtraction metricExtraction;
+
+    /**
+     * The duration of data to be searched for SAVED_SEARCH tasks,
+     * used when the task fires to calculate the query time range.
+     * <p>
+     * Duration in ISO 8601 extended format as described in
+     * https://en.wikipedia.org/wiki/ISO_8601#Durations.
+     * The value should be positive.
+     * The largest supported unit (as opposed to value) is D, e.g.  P14D (not P2W).
+     * <p>
+     * There are restrictions on the maximum duration value relative to the task schedule
+     * value as specified in the following table.
+     *    Schedule Interval Range          | Maximum Duration
+     * ----------------------------------- | -----------------
+     *   5 Minutes     to 30 Minutes       |   1 hour  \"PT60M\"
+     *  31 Minutes     to  1 Hour          |  12 hours \"PT720M\"
+     *  1 Hour+1Minute to  1 Day           |   1 day   \"P1D\"
+     *  1 Day+1Minute  to  1 Week-1Minute  |   7 days  \"P7D\"
+     *  1 Week         to  2 Weeks         |  14 days  \"P14D\"
+     *  greater than 2 Weeks               |  30 days  \"P30D\"
+     * <p>
+     * If not specified, the duration will be based on the schedule. For example,
+     * if the schedule is every 5 minutes then the savedSearchDuration will be \"PT5M\";
+     * if the schedule is every 3 weeks then the savedSearchDuration will be \"P21D\".
+     *
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("savedSearchDuration")
+    String savedSearchDuration;
 
     @com.fasterxml.jackson.annotation.JsonIgnore
     private final java.util.Set<String> __explicitlySet__ = new java.util.HashSet<String>();